CLDSRV-805: Regroup and dedup bucket Head / Get
Head tests where twice in head and get file.

Regroup in 1 file for 1 bucket for both tests kind to reduce GCP quota pressure
